use dmenu and omnihist for surfgo C-g and surffind C-/
- uses SETPROP(readprop,wyebcmd,prompt) rather than
  SETPROP(readprop,setprop,prompt)

- get redirects into omnihist's dmenu prompt via X prop

- history files are stored in ${HISTFN}_${propname}. ${HISTFN} is
a pathname prefix and  the C-g surfgo history defaults to
~/.cache/wyeb/.history_open.

- zombies are avoided by using g_subprocess_newv.

the surf sigchld(0) mechanism fails here because glib installs its
own sigchld handler - which overrides out installed handler - when the
webprocess is launched.  Use g_subprocess_newv to spawn processes
instead of a fork and exec. glib will reap the process itself.

This failure of surfexec's mechanism to reap omnihist zombies was
observed only in webkit 2.23.1. It is not clear why was not observed
in earlier versions. One observed difference was that networkprocess
was observed to be launched ahead of the webprocess in this version.
glib only installs its unix sigaction handler once for each
child. Possibly the the handler was installed before our sigchild ran.

240822
* omnihist-wyeb: use xid for wyebsuffix instead of //