w3mmode
the w3mmode setting from conf can be one of "off", "same_host", or
"one".

the UI process can override the setting from conf in the extension.
In this case the w3mmode_status slot of the Page is set a value other
than W3MMODE_USECONF.

TODO: remove network-offline which piggybacked on to this patch