main.c: extensions directory
use "WEBKIT_EXT_DIR" environment variable if present.

if DEBUG check for ext.so at the location of the binary via
/proc/self/exe. If found, use it. Otherwise check the current
directory for ext.so. If found use it. Otherwise check the
compiled-in default for ext.so. If found use it.

Otherwise fallback to wyeb's default mechanism: if DEBUG use
the current directory otherwise use the compiled-in default.