main.c: (newwin): use XEMBED environment variable to gtk_plug_new
221222
commented out. The idea was to use XEMBED always if it was present
but it would be more flexible to use plugto to put some windows
out of the tabbed window.

so instead of tabbed.sh wyeb
use a wyeb-tabbed.sh which does a
wyeb / plugto $(printf "%d" $(cat ~/tmp/tabbed-wyeb.xid))
and use wyeb / plugto 0 to send windows to send to root.