main.c: (historycb):  resize logsize if history file does not exist
When h1 h2 h3 are all > 22222 bytes, there is no h4, At the end of the loop:
 	if (!current || !g_file_test(logdir, G_FILE_TEST_EXISTS)) { .... }
`current' gets set to h4 but `logsize' is still size of h3.

logsize should be reset if file does not exist.

> Yes it is bit odd but when the size is over the max, the last
> code in the func creates a blank file.
> So when h3 reached max then there should be h4 with 0 byte.

h3 has reached max but `current' is h4 at exit from the loop
logsize is still size of h3.

now line is logged to h4, h4 is created (by `append') with only 1 line
a few bytes.

now logsize += strlen(line), logsize > 22222

so h5 is created with 0 bytes. (not h4!)
and h4 ends up with only one line

> wyeb doesen't support the case (h1 h2 h3 > max and no h4)
> because normally it should be (h1 h2 h3 > max and h4 < max)