main.c: (policyb): print stdoutheaders of the response
so far the stdoutheaders boolean option was used only in the webkit
webextension (in the web process) to print request headers (and
perhaps a redirect response). this commit intends to use the option to
print response headers: it has to be done in the main ui process.

notes: can't use static headerout