fix last commit: use webkit-user-message on the WebContext rather than the WebView
see comment in last commit.

this version uses
webkit_web_extension_send_message_to_context in ext.c:(initpage) and
viewusrmsgrcv connected to user-message-received on main.c:(newwin)
WebKitWebContext.

slightly more robust, because the user message is delivered instead of
being dropped. However.

This will also fail, and the provisional process creates the page, and
webkit doesn't know which WebKitWebPage it is part of yet,
viewusrmsgrcv picks LASTWIN, but is only correct if webkit also
assignes the provisional process to the same webview.